# docs/swagger.yaml
definitions:
  models.Date:
    properties:
      time.Time:
        type: string
    type: object
  models.Project:
    properties:
      createdAt:
        type: string
      description:
        type: string
      endDate:
        $ref: '#/definitions/models.Date'
      id:
        type: string
      lastModified:
        type: string
      name:
        type: string
      startDate:
        $ref: '#/definitions/models.Date'
      status:
        type: string
      workspace:
        $ref: '#/definitions/models.Workspace'
    type: object
...
  models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      lastModified:
        type: string
      name:
        type: string
      profilePhoto:
        type: string
      role:
        type: string
      verified:
        type: boolean
    type: object
  models.Workspace:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      lastModified:
        type: string
      name:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
info:
  contact:
    email: support@hazel.local
    name: API Support
    url: https://github.com/luckmade/hazel
  description: This is the backend API for the Hazel project management application.
  title: Hazel Project Management API
  version: "1.0"
paths:
...
  /projects:
    post:
      consumes:
      - application/json
      description: Create a new project in a workspace
      parameters:
      - description: Project info
        in: body
        name: project
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create project
      tags:
      - projects
  /projects/{id}:
    delete:
      description: Delete a project by ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete project
      tags:
      - projects
    get:
      description: Get a project by ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get project
      tags:
      - projects
    patch:
      consumes:
      - application/json
      description: Update project details
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Project update info
        in: body
        name: project
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update project
      tags:
      - projects
...
  /workspaces/{id}/projects:
    get:
      description: Get all projects for a workspace
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Project'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get projects in workspace
      tags:
      - projects
...
swagger: "2.0"